Add tests for string_find_first_nonspace
 - Assert valid then cast to avoid impl. conversion

+++ src/string_helper.h
@@ -2,8 +2,8 @@
 // Licensed under BSD-3-Clause
 // Refer to the license.txt file included in the root of the project
 
-int string_get_delimiter(const char *string, char delimiter, char **dest);
-int string_get_line(const char *string, char **dest);
+int64_t string_get_delimiter(const char *string, char delimiter, char **dest);
+int64_t string_get_line(const char *string, char **dest);
 int64_t string_find_first_nonspace(const char *string);
 int64_t string_find_last_nonspace(const char *string);
 int64_t string_strip_surrounding_space(const char *string, char **dest);
